Cleaned up, parallel correlation done.
Cleaned up some of the files, removing commented out code, adding
result to store value from fread so no warnings are emitted.

Parallel correlation calculation done using 5 value method. Results
written out to file to be loaded again (some values were outside of
range due to representation errors so were truncated).

+++ source/predictor_korbell.h
@@ -40,7 +40,7 @@ private:
 	double predict_neighbour(uint64_t user, uint64_t movie, uint64_t day);
 	static int neighbour_compare(const void *a, const void *b);
 	
-	float *correlation_intermediates;
+	double *correlation_intermediates;
 	double *residual_averages;
 	
 	typedef struct {
